Reexport Validate macro from validator crate
This allows for the reexport of the Validate macro from the root
validator crate if the `validate` feature is enabled.

Additionally, the following changes were made:
- Addition of the `validator_types` crate for the core type to break the
  cyclic dependency
- Addition of a Cargo.toml at the top-level to make the repo a workspace
- Transition from compiletest-rs to trybuild for proc macro ui testing
- Dropped the pinned test matrix due to a broken upstream dependency
- Updated various bits of code to match Rust 2018 idioms more closely

This resolves #58

+++ validator/src/validation/range.rs
@@ -32,13 +32,13 @@ mod tests {
     #[test]
     fn test_validate_range_ok() {
         let validator = Validator::Range { min: Some(0.0), max: Some(10.0) };
-        assert_eq!(validate_range(validator, 1 as f64), true);
+        assert_eq!(validate_range(validator, 1_f64), true);
     }
 
     #[test]
     fn test_validate_range_fail() {
         let validator = Validator::Range { min: Some(0.0), max: Some(10.0) };
-        assert_eq!(validate_range(validator, 20 as f64), false);
+        assert_eq!(validate_range(validator, 20_f64), false);
     }
 
     #[test]
